+++ crates/primitives/src/sol/types.rs
@@ -34,6 +34,7 @@ use ink_prelude::{
     string::String,
     vec::Vec,
 };
+use itertools::Itertools;
 use paste::paste;
 use primitive_types::U256;
 
@@ -51,12 +52,15 @@ use crate::types::Address;
 /// | `uN` for `N ∈ {8,16,32,64,128}` | `uintN` | e.g `u8` ↔ `uint8` |
 /// | `U256` | `uint256` ||
 /// | `String` | `string` ||
+/// | `Box<str>` | `string` ||
 /// | `Address` / `H160` | `address` | `Address` is a type alias for the `H160` type used for addresses in `pallet-revive` |
 /// | `[T; N]` for `const N: usize` | `T[N]` | e.g. `[i8; 64]` ↔ `int8[64]` |
 /// | `Vec<T>` | `T[]` | e.g. `Vec<i8>` ↔ `int8[]` |
-/// | `SolBytes<u8>` |  `bytes1` ||
-/// | `SolBytes<[u8; N]>` for `1 <= N <= 32` |  `bytesN` | e.g. `SolBytes<[u8; 32]>` ↔ `bytes32` |
-/// | `SolBytes<Vec<u8>>` |  `bytes` ||
+/// | `Box<[T]>` | `T[]` | e.g. `Box<[i8]>` ↔ `int8[]` |
+/// | `SolBytes<u8>` | `bytes1` ||
+/// | `SolBytes<[u8; N]>` for `1 <= N <= 32` | `bytesN` | e.g. `SolBytes<[u8; 32]>` ↔ `bytes32` |
+/// | `SolBytes<Vec<u8>>` | `bytes` ||
+/// | `SolBytes<Box<[u8]>>` | `bytes` ||
 /// | `(T1, T2, T3, ... T12)` | `(U1, U2, U3, ... U12)` | where `T1` ↔ `U1`, ... `T12` ↔ `U12` e.g. `(bool, u8, Address)` ↔ `(bool, uint8, address)` |
 ///
 /// Ref: <https://docs.soliditylang.org/en/latest/abi-spec.html#types>
@@ -73,11 +77,13 @@ pub trait SolTypeDecode: Sized + private::Sealed {
     fn decode(data: &[u8]) -> Result<Self, alloy_sol_types::Error> {
         // Don't validate decoding. Validating results in encoding and decoding again.
         abi::decode::<<Self::AlloyType as AlloySolType>::Token<'_>>(data, false)
-            .map(Self::detokenize)
+            .and_then(Self::detokenize)
     }
 
     /// Detokenizes this type's value from the given token.
-    fn detokenize(token: <Self::AlloyType as AlloySolType>::Token<'_>) -> Self;
+    fn detokenize(
+        token: <Self::AlloyType as AlloySolType>::Token<'_>,
+    ) -> Result<Self, alloy_sol_types::Error>;
 }
 
 /// A Rust/ink! equivalent of a Solidity ABI type that implements logic for Solidity ABI
@@ -92,16 +98,19 @@ pub trait SolTypeDecode: Sized + private::Sealed {
 /// | `uN` for `N ∈ {8,16,32,64,128}` | `uintN` | e.g `u8` ↔ `uint8` |
 /// | `U256` | `uint256` ||
 /// | `String` | `string` ||
+/// | `Box<str>` | `string` ||
 /// | `Address` / `H160` | `address` | `Address` is a type alias for the `H160` type used for addresses in `pallet-revive` |
 /// | `[T; N]` for `const N: usize` | `T[N]` | e.g. `[i8; 64]` ↔ `int8[64]` |
 /// | `Vec<T>` | `T[]` | e.g. `Vec<i8>` ↔ `int8[]` |
+/// | `Box<[T]>` | `T[]` | e.g. `Box<[i8]>` ↔ `int8[]` |
 /// | `SolBytes<u8>` |  `bytes1` ||
-/// | `SolBytes<[u8; N]>` for `1 <= N <= 32` |  `bytesN` | e.g. `SolBytes<[u8; 32]>` ↔ `bytes32` |
-/// | `SolBytes<Vec<u8>>` |  `bytes` ||
+/// | `SolBytes<[u8; N]>` for `1 <= N <= 32` | `bytesN` | e.g. `SolBytes<[u8; 32]>` ↔ `bytes32` |
+/// | `SolBytes<Vec<u8>>` | `bytes` ||
+/// | `SolBytes<Box<[u8]>>` | `bytes` ||
 /// | `(T1, T2, T3, ... T12)` | `(U1, U2, U3, ... U12)` | where `T1` ↔ `U1`, ... `T12` ↔ `U12` e.g. `(bool, u8, Address)` ↔ `(bool, uint8, address)` |
-/// | `&str`, `&mut str`, `Box<str>` | `string` ||
+/// | `&str`, `&mut str` | `string` ||
 /// | `&T`, `&mut T`, `Box<T>` | `T` | e.g. `&i8 ↔ int8` |
-/// | `&[T]`, `&mut [T]`, `Box<[T]>` | `T[]` | e.g. `&[i8]` ↔ `int8[]` |
+/// | `&[T]`, `&mut [T]` | `T[]` | e.g. `&[i8]` ↔ `int8[]` |
 ///
 /// Ref: <https://docs.soliditylang.org/en/latest/abi-spec.html#types>
 ///
@@ -128,8 +137,8 @@ macro_rules! impl_primitive_decode {
             impl SolTypeDecode for $ty {
                 type AlloyType = $sol_ty;
 
-                fn detokenize(token: <Self::AlloyType as AlloySolType>::Token<'_>) -> Self {
-                    <Self::AlloyType as AlloySolType>::detokenize(token)
+                fn detokenize(token: <Self::AlloyType as AlloySolType>::Token<'_>) -> Result<Self, alloy_sol_types::Error> {
+                    Ok(<Self::AlloyType as AlloySolType>::detokenize(token))
                 }
             }
         )*
@@ -184,16 +193,41 @@ impl_primitive! {
     String => sol_data::String,
 }
 
+// Rust `Box<str>` (i.e. boxed string slice) <-> Solidity `string`.
+impl SolTypeDecode for Box<str> {
+    type AlloyType = sol_data::String;
+
+    fn detokenize(
+        token: <Self::AlloyType as AlloySolType>::Token<'_>,
+    ) -> Result<Self, alloy_sol_types::Error> {
+        Ok(Box::from(core::str::from_utf8(token.0).map_err(|_| {
+            alloy_sol_types::Error::type_check_fail(token.0, "string")
+        })?))
+    }
+}
+
+impl SolTypeEncode for Box<str> {
+    type AlloyType = sol_data::String;
+
+    fn tokenize(&self) -> <Self::AlloyType as AlloySolType>::Token<'_> {
+        PackedSeqToken(self.as_bytes())
+    }
+}
+
+impl private::Sealed for Box<str> {}
+
 // Address <-> address
 impl SolTypeDecode for Address {
     type AlloyType = sol_data::Address;
 
-    fn detokenize(token: <Self::AlloyType as AlloySolType>::Token<'_>) -> Self {
+    fn detokenize(
+        token: <Self::AlloyType as AlloySolType>::Token<'_>,
+    ) -> Result<Self, alloy_sol_types::Error> {
         // We skip the conversion to `alloy_sol_types::private::Address` which will end up
         // just taking the last 20 bytes of `alloy_sol_types::abi::token::WordToken` as
         // well anyway.
         // Ref: <https://github.com/alloy-rs/core/blob/5ae4fe0b246239602c97cc5a2f2e4bc780e2024a/crates/primitives/src/bits/address.rs#L132-L134>
-        Address::from_slice(&token.0 .0[12..])
+        Ok(Address::from_slice(&token.0 .0[12..]))
     }
 }
 
@@ -216,8 +250,10 @@ impl private::Sealed for Address {}
 impl SolTypeDecode for U256 {
     type AlloyType = sol_data::Uint<256>;
 
-    fn detokenize(token: <Self::AlloyType as AlloySolType>::Token<'_>) -> Self {
-        U256::from_big_endian(token.0 .0.as_slice())
+    fn detokenize(
+        token: <Self::AlloyType as AlloySolType>::Token<'_>,
+    ) -> Result<Self, alloy_sol_types::Error> {
+        Ok(U256::from_big_endian(token.0 .0.as_slice()))
     }
 }
 
@@ -239,10 +275,25 @@ impl private::Sealed for U256 {}
 impl<T: SolTypeDecode, const N: usize> SolTypeDecode for [T; N] {
     type AlloyType = sol_data::FixedArray<T::AlloyType, N>;
 
-    fn detokenize(token: <Self::AlloyType as AlloySolType>::Token<'_>) -> Self {
+    fn detokenize(
+        token: <Self::AlloyType as AlloySolType>::Token<'_>,
+    ) -> Result<Self, alloy_sol_types::Error> {
         // Takes advantage of optimized `SolTypeDecode::detokenize` implementations and
         // skips unnecessary conversions to `T::AlloyType::RustType`.
-        token.0.map(<T as SolTypeDecode>::detokenize)
+        // FIXME: (@davidsemakula) replace with `array::try_map` if it's ever stabilized.
+        // Ref: <https://github.com/rust-lang/rust/issues/79711>
+        // Ref: <https://doc.rust-lang.org/nightly/std/primitive.array.html#method.try_map>
+        token
+            .0
+            .into_iter()
+            .map(<T as SolTypeDecode>::detokenize)
+            .process_results(|iter| iter.collect_array())?
+            .ok_or_else(|| {
+                alloy_sol_types::Error::custom(ink_prelude::format!(
+                    "ABI decoding failed: {}",
+                    Self::AlloyType::SOL_NAME
+                ))
+            })
     }
 }
 
@@ -264,7 +315,9 @@ impl<T: private::Sealed, const N: usize> private::Sealed for [T; N] {}
 impl<T: SolTypeDecode> SolTypeDecode for Vec<T> {
     type AlloyType = sol_data::Array<T::AlloyType>;
 
-    fn detokenize(token: <Self::AlloyType as AlloySolType>::Token<'_>) -> Self {
+    fn detokenize(
+        token: <Self::AlloyType as AlloySolType>::Token<'_>,
+    ) -> Result<Self, alloy_sol_types::Error> {
         // Takes advantage of optimized `SolTypeDecode::detokenize` implementations and
         // skips unnecessary conversions to `T::AlloyType::RustType`.
         token
@@ -287,6 +340,35 @@ impl<T: SolTypeEncode> SolTypeEncode for Vec<T> {
 
 impl<T: private::Sealed> private::Sealed for Vec<T> {}
 
+// Rust `Box<[T]>` (i.e. boxed slice) <-> Solidity dynamic size array (i.e. `T[]`).
+impl<T: SolTypeDecode> SolTypeDecode for Box<[T]> {
+    type AlloyType = sol_data::Array<T::AlloyType>;
+
+    fn detokenize(
+        token: <Self::AlloyType as AlloySolType>::Token<'_>,
+    ) -> Result<Self, alloy_sol_types::Error> {
+        // Takes advantage of optimized `SolTypeDecode::detokenize` implementations and
+        // skips unnecessary conversions to `T::AlloyType::RustType`.
+        token
+            .0
+            .into_iter()
+            .map(<T as SolTypeDecode>::detokenize)
+            .collect()
+    }
+}
+
+impl<T: SolTypeEncode> SolTypeEncode for Box<[T]> {
+    type AlloyType = sol_data::Array<T::AlloyType>;
+
+    fn tokenize(&self) -> <Self::AlloyType as AlloySolType>::Token<'_> {
+        // Does NOT require `SolValueType<Self::AlloyType>` and instead relies on
+        // `SolTypeEncode::tokenize`.
+        DynSeqToken(self.iter().map(<T as SolTypeEncode>::tokenize).collect())
+    }
+}
+
+impl<T: private::Sealed> private::Sealed for Box<[T]> {}
+
 // We follow the Rust standard library's convention of implementing traits for tuples up
 // to twelve items long.
 // Ref: <https://doc.rust-lang.org/std/primitive.tuple.html#trait-implementations>
@@ -295,10 +377,12 @@ impl SolTypeDecode for Tuple {
     for_tuples!( type AlloyType = ( #( Tuple::AlloyType ),* ); );
 
     #[allow(clippy::unused_unit)]
-    fn detokenize(token: <Self::AlloyType as AlloySolType>::Token<'_>) -> Self {
+    fn detokenize(
+        token: <Self::AlloyType as AlloySolType>::Token<'_>,
+    ) -> Result<Self, alloy_sol_types::Error> {
         // Takes advantage of optimized `SolTypeDecode::detokenize` implementations and
         // skips unnecessary conversions to `T::AlloyType::RustType`.
-        for_tuples!( ( #( <Tuple as SolTypeDecode>::detokenize(token.Tuple) ),* ) );
+        Ok(for_tuples! { ( #( <Tuple as SolTypeDecode>::detokenize(token.Tuple)? ),* ) })
     }
 }
 
@@ -368,7 +452,7 @@ macro_rules! impl_str_ref_encode {
     };
 }
 
-impl_str_ref_encode!(&str, &mut str, Box<str>);
+impl_str_ref_encode!(&str, &mut str);
 
 macro_rules! impl_slice_ref_encode {
     ($($ty: ty),+ $(,)*) => {
@@ -388,7 +472,7 @@ macro_rules! impl_slice_ref_encode {
     };
 }
 
-impl_slice_ref_encode!(&[T], &mut [T], Box<[T]>);
+impl_slice_ref_encode!(&[T], &mut [T]);
 
 pub(super) mod private {
     /// Seals implementations of `SolTypeEncode` and `SolTypeDecode`.